refactor package; use avro specific records (vs generic); change topic
files moved to subfolder for maven support.

the topic should really be in a config file, but this is fine for now.

avro specific records should allow an easier deserialization workflow. I'm not sure I'm doing it efficiently yet, but I think it works. I'll work out the kinks as I can, but demo is in 2 days and there's still a lot of functionality I need to implement. Technical debt.