add venturi project: the kafka stream translator / filter
This project is intended to run on kafka nodes, ingesting fat avro from the github "firehose", emitting filtered avro records for consumption downstream. I haven't worked out distribution yet, I believe kafka streams will handle a lot of it, but it's definitely a TODO. For Friday, I need to have and end-to-end working, I'll have to scale this later.