flinkCC: implement event-time windowing
Unfortunately, the data exported from BigQuery is not sorted by event
time. Sorting it right now, given everything else that must be
finished, would be herculean and probably foolish.

The event-time logic is there and apparently functional. If I have
time to build a pipeline to sort 1.2TB of avro data, I will.