flinkCC: add restart strategy and go back to processing time windows
Event time windows were a bust due to massively out-of-order data
(months off).

As it turns out, flink streams have an optional built-in restart
strategy. This is the watchdog I've been looking for. There are still
occasions when kafka goes through a leader change, which cause the
flink-kafka connector to crash (intentionally, if I am to believe the
jira ticket). Also, if flinkCC crashes for a bit, it's possible to
overload a window and cause a ~1MB message to try to pass through
kafka, which is rejected, and this also crashes the flink job. The
restart strategy appears to take care of that. 99999 retries with 0
wait.