firehose: deal with changing event schema
Ideally I'd update the single avro schema, but this isn't simply a set
of column additions. For whatever reason, the avro schema for the 2015
files is not nested like the 2011-2014 data. The BigQuery column names
look identical between the years, I'm not sure why BigQuery exported
the 2015 in a flat format. Regardless, this seems like a fine
workaround.